[udev][usb-xhci] Remove an unnecessary malloc from iotxn_queue
We used to malloc a xhci_transfer_context_t struct for every USB transaction
that was queued. Now we use the iotxn proto_data space for this instead.
Since the USB bus driver is now cloning transactions passed from above,
the upper level drivers will not be affected by this.

Some small changes to how deferred transactions were necessary
in order to get this to work.

Change-Id: Iecb1589c961191e02c6029d4c28cb653e9521f61