[udev][usb-mass-storage] Respect USB controller's max transfer size
If we receive a transfer that exceeds the max transfer size reported by
the USB controller driver, we now process the data phase in multiple chunks
by repeatedly requeueing the iotxn, adjusting the vmo_offset and length
as necessary.

This change is to prepare for the new BLOCK_CORE protocol, which will be
giving us arbitrarily large blocks of data to transfer.

Change-Id: I1d5593c7273c0eacb7cfe70551d6aa6bc8c2e1b6